+++ R/rbind_viterbi.R
@@ -8,7 +8,7 @@
 #'
 #' @return An object of class `"viterbi"`, like the input; see [viterbi()].
 #'
-#' @seealso [cbind.viterbi()], [viterbi()]
+#' @seealso [cbind.viterbi()], [subset.viterbi()], [viterbi()]
 #'
 #' @examples
 #' grav2 <- read_cross2(system.file("extdata", "grav2.zip", package="qtl2"))